refactored types into a package
