+++ starlite/types/composite.py
@@ -1,0 +1,35 @@
+from typing import TYPE_CHECKING, Any, Callable, Dict, List, Type, Union
+
+from .asgi_types import ASGIApp
+from .callable_types import ExceptionHandler
+
+if TYPE_CHECKING:
+    from pydantic.fields import FieldInfo  # noqa: TC004
+    from starlette.middleware import Middleware as StarletteMiddleware  # noqa: TC004
+    from starlette.middleware.base import BaseHTTPMiddleware  # noqa: TC004
+
+    from starlite.datastructures import Cookie, ResponseHeader  # noqa: TC004
+    from starlite.middleware.base import (  # noqa: TC004
+        DefineMiddleware,
+        MiddlewareProtocol,
+    )
+    from starlite.provide import Provide  # noqa: TC004
+else:
+    BaseHTTPMiddleware = Any
+    Cookie = Any
+    DefineMiddleware = Any
+    FieldInfo = Any
+    MiddlewareProtocol = Any
+    Provide = Any
+    ResponseHeader = Any
+    StarletteMiddleware = Any
+
+Dependencies = Dict[str, Provide]
+ExceptionHandlersMap = Dict[Union[int, Type[Exception]], ExceptionHandler]
+
+Middleware = Union[
+    Callable[..., ASGIApp], DefineMiddleware, StarletteMiddleware, Type[BaseHTTPMiddleware], Type[MiddlewareProtocol]
+]
+ParametersMap = Dict[str, FieldInfo]
+ResponseCookies = List[Cookie]
+ResponseHeadersMap = Dict[str, ResponseHeader]
