+++ starlite/types/empty.py
@@ -1,0 +1,8 @@
+from typing import Type
+
+
+class Empty:
+    """A sentinel class used as placeholder."""
+
+
+EmptyType = Type[Empty]
