+++ starlite/types/helpers.py
@@ -1,0 +1,6 @@
+from typing import Awaitable, List, TypeVar, Union
+
+T = TypeVar("T")
+
+SyncOrAsyncUnion = Union[T, Awaitable[T]]
+SingleOrList = Union[T, List[T]]
